+++ pynfe/utils/flags.py
@@ -10,6 +10,11 @@ NAMESPACE_METODO = 'http://www.portalfiscal.inf.br/nfe/wsdl/'
 NAMESPACE_SOAP_NFSE = 'http://schemas.xmlsoap.org/soap/envelope/'
 NAMESPACE_BETHA = 'http://www.betha.com.br/e-nota-contribuinte-ws'
 
+NAMESPACE_MDFE = 'http://www.portalfiscal.inf.br/mdfe'
+NAMESPACE_MDFE_METODO = 'http://www.portalfiscal.inf.br/mdfe/wsdl/'
+MODELO_MDFE = '58'
+VERSAO_MDFE = '3.00'
+
 VERSAO_PADRAO = '4.00'
 
 VERSAO_QRCODE = '2'
